+++ src/ui/ModListItem.hpp
@@ -5,17 +5,18 @@
 using namespace geode::prelude;
 
 // Mod list item
-class ModItem : public CCNode {
+class ModListItem : public CCNode {
 private:
     class Impl;
     std::unique_ptr<Impl> m_impl;
 
 protected:
-    ModItem();
-    virtual ~ModItem();
+    ModListItem();
+    virtual ~ModListItem();
 
     void onViewMod(CCObject*);
     void onModDesc(CCObject*);
+
     void onFavorite(CCObject*);
 
     CCLabelBMFont* firstTimeText();
@@ -24,5 +25,5 @@ protected:
     bool init(Mod* mod, CCSize const& size, bool geodeTheme, bool heartIcons);
 
 public:
-    static ModItem* create(Mod* mod, CCSize const& size, bool geodeTheme = false, bool heartIcons = false);
+    static ModListItem* create(Mod* mod, CCSize const& size, bool geodeTheme = false, bool heartIcons = false);
 };